+++ cogs/servercommands.py
@@ -33,9 +33,10 @@ class ServerCommands(commands.Cog):
         self.bot = bot
 
     @slash_command(name="addstaff")
-    @permissions.has_role("Bot Manager")
     async def addstaff(self, ctx, role: discord.Role):
-        """Adds a role to the staff list"""
+        """[STAFF] Adds a role to the staff list"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
         if role.id in GG.STAFF:
             return await ctx.respond(f"{role.name} is already on the staff list.", ephemeral=True)
         else:
@@ -44,9 +45,10 @@ class ServerCommands(commands.Cog):
             return await ctx.respond(f"{role.name} was added to the staff list.", ephemeral=True)
 
     @slash_command(name="stafflist")
-    @permissions.has_role("Bot Manager")
     async def stafflist(self, ctx):
-        """Returns a list of all staff roles"""
+        """[STAFF] Returns a list of all staff roles"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
         user_quotes = await GG.MDB['serverstaff'].find({"guild": ctx.guild.id}).to_list(length=None)
         if user_quotes is not None:
             return await ctx.respond(embed=list_embed(user_quotes, ctx.author, ctx), ephemeral=True)
@@ -54,9 +56,10 @@ class ServerCommands(commands.Cog):
             return await ctx.respond("This server has no roles assigned as staff.", ephemeral=True)
 
     @slash_command(name="delstaff")
-    @permissions.has_role("Bot Manager")
     async def delstaff(self, ctx, role: discord.Role):
-        """Removes a role to the staff list"""
+        """[STAFF] Removes a role to the staff list"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
         exist = await GG.MDB['serverstaff'].find_one({"guild": ctx.guild.id, "roles": int(role.id)})
         if exist is not None:
             await GG.MDB['serverstaff'].delete_one({"guild": ctx.guild.id, "roles": int(role.id)})
@@ -66,9 +69,10 @@ class ServerCommands(commands.Cog):
             return await ctx.respond(f"{role.name} is not on the staff list.", ephemeral=True)
 
     @slash_command(name="prefix")
-    @permissions.has_role("Bot Manager")
     async def prefix(self, ctx, prefix: Option(str, "What do you want the prefix to be")):
-        """Sets the bot's prefix for this server. Forgot the prefix? Reset it with "/prefix $"""
+        """[STAFF] Sets the bot's prefix for this server. Forgot the prefix? Reset it with "/prefix $"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
         guild_id = str(ctx.guild.id)
         if prefix is None:
             current_prefix = await self.bot.get_server_prefix(ctx.message)
@@ -85,9 +89,10 @@ class ServerCommands(commands.Cog):
         return await ctx.respond("Prefix set to `{}` for this server.".format(prefix), ephemeral=True)
 
     @slash_command(name="addchannel")
-    @permissions.has_role("Bot Manager")
     async def addchannel(self, ctx, channeltype: Option(str, "Choose your channel type", choices=categories), channel: Option(discord.TextChannel, "For which channel do you want to set the type?")):
-        """Gives a channel a specific type."""
+        """[STAFF] Gives a channel a specific type."""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
         channelExists = await GG.MDB['channelinfo'].find_one({"channel": channel.id})
         if channelExists is not None:
             return await ctx.respond(f"{channel.name} is already on the special channel list.", ephemeral=True)
@@ -97,9 +102,10 @@ class ServerCommands(commands.Cog):
             return await ctx.respond(f"{channel.name} was added to the special channel list as {channeltype}", ephemeral=True)
 
     @slash_command(name="delchannel")
-    @permissions.has_role("Bot Manager")
     async def delchannel(self, ctx, channel: Option(discord.TextChannel, "From what channel do you want to remove the channel type?")):
-        """Removes the type of a channel, reverting it to a "normal" channel"""
+        """[STAFF] Removes the type of a channel, reverting it to a "normal" channel"""
+        if not GG.is_staff_bool(ctx):
+            return await ctx.respond("You do not have the required permissions to use this command.", ephemeral=True)
         channelExists = await GG.MDB['channelinfo'].find_one({"channel": channel.id})
         if channelExists is not None:
             await GG.MDB['channelinfo'].delete_one({"channel": channel.id})
